Update Utils.java
If you try to use the version 1.01 and try to get the Quotes, the method called parseDividendDate is giving an parsing exception. 
Reason:- When trying for the symbol "JCP" and extra space is present in the date and thus the above method fails. Adding trim method in here should fix the problem. Please test it out.